requirements.txt
# Inference
exllamav2>=0.3.2

# HTTP client
requests>=2.32.0

# Document ingestion
PyMuPDF>=1.24.0            # PDF text/table extraction
python-docx>=1.1.0         # DOCX parsing
openpyxl>=3.1.0            # Excel file support
pandas>=2.2.0              # CSV/Excel data handling
bibtexparser>=2.0.0        # BibTeX reference parsing

# Text processing
tiktoken>=0.7.0            # Token counting
beautifulsoup4>=4.12.0     # HTML cleaning

# Embeddings (CPU-only)
sentence-transformers>=3.0.0

# Vector DB for RAG
chromadb>=0.5.0

# Literature search
arxiv>=2.1.0               # arXiv API client
biopython>=1.83            # PubMed/NCBI E-utilities

# Web research
duckduckgo-search>=6.0.0
trafilatura>=1.12.0

# Web dashboard
flask>=3.0.0

# Config
pyyaml>=6.0.0

# TTS
edge-tts>=6.1.0
